Build parent-child mapping on load
Build parent-child mapping so that we can create parent IDs on proxy
items where the parent contains a collection and there is no item on the
child.  E.g. given Author.Books, Author is the parent and Book is the
child and there should be an AuthorID property created on the Book proxy
class